server/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pandas==2.2.0
python-dotenv==1.0.1
runeq==0.17.2
pydantic==2.6.1
python-multipart==0.0.9 